# Conflicts:
#	src/main/java/com/higgsup/kpi/service/impl/EventServiceImpl.java

+++ database/versions/2018-10-25-1119_v15.sql
@@ -223,7 +223,7 @@ CREATE TABLE `kpi_point_detail` (
   `user_name`             VARCHAR(50)
                           CHARACTER SET utf8 NOT NULL,
   `point_type`            INT(11)            NOT NULL,
-  `point`                 INT(11)            NOT NULL,
+  `point`                 FLOAT              NOT NULL,
   `year_month_id`         INT(11)            NULL,
   FOREIGN KEY (event_id) REFERENCES kpi_event (id),
   FOREIGN KEY (user_name) REFERENCES kpi_user (user_name),
